+++ src/components/GanttRoot.vue
@@ -30,6 +30,7 @@ import { DEFAULT_ZOOM_LEVELS } from '../zoom'
 import type {
   GanttBand,
   GanttCellEvent,
+  GanttCreateEvent,
   GanttColumn,
   GanttColumnEvent,
   GanttConfig,
@@ -86,6 +87,7 @@ const props = withDefaults(defineProps<GanttRootProps>(), {
   criticalPath: GANTT_DEFAULTS.criticalPath,
   slack: GANTT_DEFAULTS.slack,
   linkable: GANTT_DEFAULTS.linkable,
+  cellCreatable: GANTT_DEFAULTS.cellCreatable,
   dependencyShape: elbowPath,
   arrowHead: triangleArrow,
   snapToGrid: GANTT_DEFAULTS.snapToGrid,
@@ -135,6 +137,7 @@ const emit = defineEmits<{
   'row-contextmenu': [event: GanttRowEvent]
   'cell-click': [event: GanttCellEvent]
   'cell-dblclick': [event: GanttCellEvent]
+  'create': [event: GanttCreateEvent]
   'column-click': [event: GanttColumnEvent]
   'dependency-click': [event: GanttDependencyEvent]
 }>()
@@ -378,6 +381,7 @@ const config = computed<GanttConfig>(() => ({
   criticalPath: props.criticalPath,
   slack: props.slack,
   linkable: props.linkable,
+  cellCreatable: props.cellCreatable,
   dependencyShape: props.dependencyShape,
   arrowHead: props.arrowHead,
   snapToGrid: props.snapToGrid,